requirements.txt
certifi==2018.4.16
chardet==3.0.4
crc16==0.1.1
defusedxml==0.5.0
Django==2.0.4
django-braces==1.13.0
django-oauth-toolkit==1.1.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
idna==2.6
oauthlib==2.0.7
PyJWT==1.6.1
python-redis==0.1.7
python3-openid==3.1.0
pytz==2018.4
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
urllib3==1.22
